requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
praw==5.2.0
prawcore==0.12.0
python-dotenv==0.7.1
requests==2.18.4
requests-file==1.4.2
six==1.11.0
tldextract==2.2.0
update-checker==0.16
urllib3==1.22
